pyproject.toml
[project]
name = "p3394-agent-starter"
version = "0.2.0"
description = "P3394 Agent Starter Kit - Build P3394-compliant agents with this template"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "P3394 Community"}
]
keywords = ["agent", "ai", "p3394", "llm", "starter-kit", "template"]

dependencies = [
    "claude-agent-sdk>=0.1.20",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
    "anyio>=4.0.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
    "qrcode>=7.4.0",
    "jinja2>=3.1.6",
    "cryptography>=46.0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
p3394-agent = "p3394_agent.__main__:run"
p3394-cli = "p3394_agent.cli_client:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/p3394_agent", "config"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
